réparation de bullshit

+++ src/game/artefacts/Artefact.java
@@ -16,9 +16,9 @@ public class Artefact implements IArtefact {
 	private int moral;
 	private int chance;
 	private int experience;
-	private double ptDeplacement;
+	private int ptDeplacement;
 
-	public Artefact(String type, int attaque, int defense, int magieMin, int magieMax, int puissanceMagie, int connaissance, int moral, int chance, int experience, double ptDeplacement) {
+	public Artefact(String type, int attaque, int defense, int magieMin, int magieMax, int puissanceMagie, int connaissance, int moral, int chance, int experience, int ptDeplacement) {
 		this.type = type;
 		this.attaque = attaque;
 		this.defense = defense;
